+++ frontend/src/App.js
@@ -12,7 +12,7 @@ import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 import ProfileTabs from "./components/profile/profileTabs";
 import PageContainer from "./components/Container";
 import JourneyEdit from "./components/JourneyEdit";
-import OrderAdd from "./components/OrderAdd";
+import OrderAdd from "./components/orderAdd/OrderAdd";
 import "antd/dist/antd.css";
 
 class App extends React.Component {
